create app folder and move calculator.py into folder
