+++ requirements.txt
@@ -5,4 +5,4 @@ packaging==23.1
 pluggy==1.0.0
 pytest==7.3.1
 tomli==2.0.1
-pytest-cov==4.1.0
+pytest-cov==4.0.0
