+++ tests/test_calculator.py
@@ -1,4 +1,4 @@
-from calculator import add, div, mul, sub
+from app.calculator import add, div, mul, sub
 
 
 def test_add():
